backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.30.0
pgvector==0.3.6
psycopg2-binary==2.9.10
alembic==1.14.0
httpx==0.28.0
openai==1.60.0
anthropic==0.43.0
pydantic==2.10.0
python-dotenv==1.0.1
tenacity==9.0.0
tqdm==4.67.0
